Fix regex to not hang on invalid input
If a user would enter a message that had the starting ``` but not the
endings ones, the regex could hang and "never" exit. Rewriting to be
smarter and trying to do less fixes that.
It now doesn't try to not capture whitespace, since is quite easy to
remove that after the fact.
This should also be faster.

Tweak the after regex check to also catch and error on empty strings.