Add a check for the compiler returning no output
If there was no output, the bot would return an empty message, which
would confuse the user. Check for an empty message, and let the user
know that their code produced no output.
This is acceptable since a bug or error in the code would have output in
the form of the error, so this simple message works for code that does
not error, but also doesn't output anything.