Expand docs and add Docker image env vars
Add environment var support to the Docker image, so that users don't
have to mount a settings.py file to the image. Also specify defaults for
the vars that make sense.
Add more information to docs, including how to use the docker image and
details of all the env vars.